third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/grid-inline-end-items-in-last-column-trimmed-expected.txt

This is a testharness.js-based test.
[FAIL] grid > item 2
  assert_equals: \n<item data-expected-margin-right="0"></item>\nmargin-right expected "0" but got "116.094"
[FAIL] grid > item 3
  assert_equals: \n<item class="locked-position" data-expected-margin-right="0"></item>\nmargin-right expected "0" but got "-30"
Harness: the test ran to completion.
